Add base napari menu class to perform action and clean up operations
