+++ napari/_qt/qt_main_window.py
@@ -1207,7 +1207,11 @@ class Window:
         _themes.events.added.disconnect(register_napari_themes)
         _themes.events.removed.disconnect(self._remove_theme)
         self.qt_viewer.viewer.layers.events.disconnect(self.file_menu.update)
-        self.file_menu.close()
+        for menu in self.file_menu._INSTANCES:
+            try:
+                menu.close()
+            except RuntimeError:
+                pass
 
     def close(self):
         """Close the viewer window and cleanup sub-widgets."""
